Rename to StreamTable and make public
